add get route that sends price data for given currencies

+++ migrations/20180425131408_price.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('price', function(table){
     table.increments();
     table.integer('currency_id').references('id').inTable('currency').notNullable();
-    table.integer('price').notNullable();
+    table.decimal('price').notNullable();
     table.timestamp('time').notNullable().defaultTo(knex.fn.now());
   });
 };
